Fixed Firefox/Edge focus outline bug & focus alignment
Applied a fix for Firefox/Edge where the css "outline-color" property doesn't draw an outline. Applied a border for those cases.

Added an alignment fix to prevent input elements breaking the page alignment when focused. This happened because of the Firefox/Edge fix I implemented, as described above. Woo! Fun!